+++ core/src/main/java/io/kestra/core/models/triggers/StatefulTriggerInterface.java
@@ -1,95 +1,40 @@
 package io.kestra.core.models.triggers;
 
-import com.fasterxml.jackson.core.type.TypeReference;
-import io.kestra.core.runners.RunContext;
-import io.kestra.core.storages.kv.*;
-import io.kestra.core.serializers.JacksonMapper;
+import io.kestra.core.models.property.Property;
+import io.swagger.v3.oas.annotations.media.Schema;
 
-import java.time.*;
-import java.util.*;
-import java.util.stream.Collectors;
+import java.time.Duration;
 
 public interface StatefulTriggerInterface {
-    record Entry(String uri, String version, Instant modifiedAt, Instant lastSeenAt) {
-        public static Entry candidate(String uri, String version, Instant modifiedAt) {
-            return new Entry(uri, version, modifiedAt, null);
-        }
-    }
+    @Schema(
+        title = "Trigger event type",
+        description = """
+            Defines when the trigger fires.
+            - `CREATE`: only for newly discovered entities.
+            - `UPDATE`: only when an already-seen entity changes.
+            - `CREATE_OR_UPDATE`: fires on either event.
+            """
+    )
+    Property<On> getOn();
+
+    @Schema(
+        title = "State key",
+        description = """
+            JSON-type KV key for persisted state.
+            Default: `<namespace>__<flowId>__<triggerId>`
+            """
+    )
+    Property<String> getStateKey();
+
+    @Schema(
+        title = "State TTL",
+        description = "TTL for persisted state entries (e.g., PT24H, P7D)."
+    )
+    Property<Duration> getStateTtl();
 
     enum On {
         CREATE,
         UPDATE,
         CREATE_OR_UPDATE
-    }
-
-    record StateUpdate(boolean fire, boolean isNew) {}
-
-    default Map<String, Entry> readState(RunContext runContext, String key, Optional<Duration> ttl) {
-        try {
-            var kv = runContext.namespaceKv(runContext.flowInfo().namespace()).getValue(key);
-            if (kv.isEmpty()) {
-                return new HashMap<>();
-            }
-
-            var entries = JacksonMapper.ofJson().readValue((byte[]) kv.get().value(), new TypeReference<List<Entry>>() {});
-
-            var cutoff = ttl.map(d -> Instant.now().minus(d)).orElse(Instant.MIN);
-
-            return entries.stream()
-                .filter(e -> Optional.ofNullable(e.lastSeenAt()).orElse(Instant.now()).isAfter(cutoff))
-                .collect(Collectors.toMap(Entry::uri, e -> e));
-        } catch (Exception e) {
-            runContext.logger().warn("readState failed", e);
-            return new HashMap<>();
-        }
-    }
-
-    default void writeState(RunContext runContext, String key, Map<String, Entry> state, Optional<Duration> ttl) {
-        try {
-            var bytes = JacksonMapper.ofJson().writeValueAsBytes(state.values());
-            var meta = new KVMetadata("trigger state", ttl.orElse(null));
-
-            runContext.namespaceKv(runContext.flowInfo().namespace()).put(key, new KVValueAndMetadata(meta, bytes));
-        } catch (Exception e) {
-            runContext.logger().warn("writeState failed", e);
-        }
-    }
-
-    default StateUpdate computeAndUpdateState(Map<String, Entry> state, Entry candidate, On on) {
-        var prev = state.get(candidate.uri());
-        var isNew = prev == null;
-        var fire = shouldFire(prev, candidate.version(), on);
-
-        Instant lastSeenAt;
-
-        if (fire || isNew) {
-            // it is new seen or changed
-            lastSeenAt = Instant.now();
-        } else if (prev.lastSeenAt() != null) {
-            // it is unchanged but already tracked before
-            lastSeenAt = prev.lastSeenAt();
-        } else {
-            lastSeenAt = Instant.now();
-        }
-
-        var newEntry = new Entry(candidate.uri(), candidate.version(), candidate.modifiedAt(), lastSeenAt);
-
-        state.put(candidate.uri(), newEntry);
-
-        return new StateUpdate(fire, isNew);
-    }
-
-    default boolean shouldFire(Entry prev, String version, On on) {
-        if (prev == null) {
-            return on == On.CREATE || on == On.CREATE_OR_UPDATE;
-        }
-        if (!Objects.equals(prev.version(), version)) {
-            return on == On.UPDATE || on == On.CREATE_OR_UPDATE;
-        }
-        return false;
-    }
-
-    default String defaultKey(String ns, String flowId, String triggerId) {
-        return String.join("_", ns, flowId, triggerId);
     }
 }
